+++ breeze-dsl/breeze-dsl-cmd-text/src/main/kotlin/CmdTextDslExtensions.kt
@@ -15,56 +15,75 @@ inline fun cmdTextDsl(block: Document.() -> CharSequence): Document {
 	return Document().apply { text = block() }
 }
 
+
 /**
  * 创建一个[CmdTextDsl.BoldText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.bold(text: CharSequence): BoldText = BoldText(text)
+fun InlineDslEntry.bold(text: CharSequence): BoldText {
+	return BoldText(text)
+}
 
 /**
  * 创建一个[CmdTextDsl.LightText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.light(text: CharSequence): LightText = LightText(text)
+fun InlineDslEntry.light(text: CharSequence): LightText {
+	return LightText(text)
+}
 
 /**
  * 创建一个[CmdTextDsl.ItalicText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.italic(text: CharSequence): ItalicText = ItalicText(text)
+fun InlineDslEntry.italic(text: CharSequence): ItalicText {
+	return ItalicText(text)
+}
 
 /**
  * 创建一个[CmdTextDsl.UnderlineText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.underline(text: CharSequence): UnderlineText = UnderlineText(text)
+fun InlineDslEntry.underline(text: CharSequence): UnderlineText {
+	return UnderlineText(text)
+}
 
 /**
  * 创建一个[CmdTextDsl.BlinkText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.blink(text: CharSequence): BlinkText = BlinkText(text)
+fun InlineDslEntry.blink(text: CharSequence): BlinkText {
+	return BlinkText(text)
+}
 
 /**
  * 创建一个[CmdTextDsl.InvertText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.invert(text: CharSequence): InvertText = InvertText(text)
+fun InlineDslEntry.invert(text: CharSequence): InvertText {
+	return InvertText(text)
+}
 
 /**
  * 创建一个[CmdTextDsl.ColoredText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.color(text: CharSequence, color: Color): ColoredText = ColoredText(text, color)
+fun InlineDslEntry.color(text: CharSequence, color: Color): ColoredText {
+	return ColoredText(text, color)
+}
 
 /**
  * 创建一个[CmdTextDsl.BgColoredText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.bgColor(text: CharSequence, color: Color): BgColoredText = BgColoredText(text, color)
+fun InlineDslEntry.bgColor(text: CharSequence, color: Color): BgColoredText {
+	return BgColoredText(text, color)
+}
 
 /**
  * 创建一个[CmdTextDsl.StyledText]。
  */
 @CmdTextDslMarker
-fun InlineDslEntry.style(text: CharSequence, vararg styles: Style): StyledText = StyledText(text, *styles)
+fun InlineDslEntry.style(text: CharSequence, vararg styles: Style): StyledText {
+	return StyledText(text, *styles)
+}