+++ breeze-dsl/breeze-dsl-critic-markup/src/main/kotlin/CriticMarkupDslExtensions.kt
@@ -15,32 +15,43 @@ inline fun criticMarkupDsl(block: Document.() -> CharSequence): Document {
 	return Document().apply { text = block() }
 }
 
+
 /**
  * 创建一个[CriticMarkupDsl.Addition]。
  */
 @CriticMarkupDslMarker
-fun InlineDslEntry.append(text: CharSequence): Addition = Addition(text)
+fun InlineDslEntry.append(text: CharSequence): Addition {
+	return Addition(text)
+}
 
 /**
  * 创建一个[CriticMarkupDsl.Deletion]。
  */
 @CriticMarkupDslMarker
-fun InlineDslEntry.delete(text: CharSequence): Deletion = Deletion(text)
+fun InlineDslEntry.delete(text: CharSequence): Deletion {
+	return Deletion(text)
+}
 
 /**
  * 创建一个[CriticMarkupDsl.Substitution]。
  */
 @CriticMarkupDslMarker
-fun InlineDslEntry.substitute(text: CharSequence, newText: CharSequence): Substitution = Substitution(text, newText)
+fun InlineDslEntry.substitute(text: CharSequence, newText: CharSequence): Substitution {
+	return Substitution(text, newText)
+}
 
 /**
  * 创建一个[CriticMarkupDsl.Comment]。
  */
 @CriticMarkupDslMarker
-fun InlineDslEntry.comment(text: CharSequence): Comment = Comment(text)
+fun InlineDslEntry.comment(text: CharSequence): Comment {
+	return Comment(text)
+}
 
 /**
  * 创建一个[CriticMarkupDsl.Highlight]。
  */
 @CriticMarkupDslMarker
-fun InlineDslEntry.highlight(text: CharSequence): Highlight = Highlight(text)
+fun InlineDslEntry.highlight(text: CharSequence): Highlight {
+	return Highlight(text)
+}