+++ breeze-dsl/breeze-dsl-critic-markup/src/main/kotlin/CriticMarkupDslMarker.kt
@@ -5,8 +5,6 @@ package com.windea.breezeframework.dsl.criticmarkup
 
 /**
  * [CriticMarkupDsl]的标记注解。参见：[Critic Markup](http://criticmarkup.com)
- *
- * Marker annotation of [CriticMarkupDsl]. See: [Critic Markup](http://criticmarkup.com)
  */
 @DslMarker
 @MustBeDocumented