+++ breeze-dsl/breeze-dsl-xml/src/main/kotlin/XmlDslMarker.kt
@@ -1,16 +1,11 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-@file:Suppress("unused", "NON_PUBLIC_PRIMARY_CONSTRUCTOR_OF_INLINE_CLASS", "INLINE_CLASS_NOT_TOP_LEVEL")
-
 package com.windea.breezeframework.dsl.xml
 
 /**
  * [XmlDsl]的标记注解。
- *
- * Marker annotation of [XmlDsl].
  */
 @DslMarker
 @MustBeDocumented
 annotation class XmlDslMarker
-