+++ breeze-dsl/src/main/kotlin/DslElement.kt
@@ -14,3 +14,4 @@ import com.windea.breezeframework.dsl.api.*
 interface DslElement {
 	override fun toString(): String
 }
+