+++ breeze-dsl/src/main/kotlin/api/DslApiMarker.kt
@@ -4,7 +4,7 @@
 package com.windea.breezeframework.dsl.api
 
 /**
- * [Breeze DslDocument Api](https://github.com/DragonKnightOfBreeze/Breeze-Framework/tree/master/breeze-dsl)的标记注解。
+ * [Breeze Dsl Api](https://github.com/DragonKnightOfBreeze/Breeze-Framework/tree/master/breeze-dsl)的标记注解。
  *
  * Marker annotation of [Breeze DslDocument Api](https://github.com/DragonKnightOfBreeze/Breeze-Framework/tree/master/breeze-dsl).
  */