+++ CHANGELOG.md
@@ -442,10 +442,18 @@
 * [X] breeze-serialization 完善`MapSerializer`
 * [X] breeze-serialization 为某一数据类型的`Serializer`如`JsonSerializer`提供作为伴生对象的默认实现。
 * [X] breeze-core 将`Any.expand`改为`List<T>.expand`
+* [X] 移除`breeze-mapper`，`breeze-serializer`，`breeze-generator`并合并到`breeze-unstable`
 * [ ] breeze-serialization 实现默认的用于处理CSV的`BreezeCsvMapper`
 * [ ] breeze-serialization 实现默认的用于处理TSV的`BreezeTsvMapper`
 * [ ] breeze-dsl 重构代码
-* [X] 移除`breeze-mapper`，`breeze-serializer`，`breeze-generator`并合并到`breeze-unstable`
+  * [X] breeze-dsl-xml
+  * [X] breeze-dsl-cmd-text
+  * [X] breeze-dsl-critic-markup
+  * [X] breeze-dsl-creole
+  * [ ] breeze-dsl-markdown
+  * [ ] breeze-dsl-flow
+  * [ ] breeze-dsl-sequence
+  * [ ] breeze-dsl-mermaid
 
 # 长期